Added methods to classes -Coach -Day -Grade -Learner -Lesson -Timeslot
Created logic for seeding timeslots and days.
-Globals
-HatfieldJuniorSwimmingSchool
-Seeder

Created a method for coloured console out.
-Utils

Created a bunch of stuff to have dual verification points at input and for each constructor. This should mean that if wrong seed data is input the program will not compile, but the program should ensure the data is valid at point of user input so no runtime errors.
-Created a validation result class in Utils to handle this

Created 2 input Systems
-user inputs int to select from menu options
-user inputs text until correct
-Both pass the validators as parameters
-Needed to do some fancy stuff to get around BiFunc and Func :), stupid java, should have used objects but that is just so messy.

Created reusable validators which can be used for user input and class construction.
		input
			ValidateAge
			ValidateContact
			ValidateGender
			ValidateGrade
			ValidateName
			ValidatePhoneNumber
		menu
			ValidateMainMenu
		others
			ValidateDay